qcacld-2.0: Fix au_add_group issues
1. group_id is passed in type u32 by both IOCTL au_add_group and
WMI command WMI_AUDIO_AGGR_ADD_GROUP_CMDID. Change its type to
u32 in function wma_add_multicast_group().
2. When assigned group_id is used or no more available group_id
in au_add_group command, put string "err" in command output and
return 0 (otherwise output can't display).

This feature should work with AUDIO_MULTICAST_AGGR_SUPPORT enabled.

Tested HW: QCA9377
CRs-Fixed: 2592825

+++ CORE/SERVICES/WMA/wma.c
@@ -17714,7 +17714,7 @@ int wma_add_multicast_group(void *wmapvosContext, int vdev_id,
 	ol_txrx_vdev_handle vdev = NULL;
 	struct ol_audio_multicast_aggr_conf* au_mcast_conf = NULL;
 	struct ol_audio_multicast_group* pMultiGroup = NULL;
-	u_int8_t group_id, group_index;
+	u_int32_t group_id, group_index;
 	tp_wma_handle wma;
 
 	wma = (tp_wma_handle) vos_get_context(VOS_MODULE_ID_WDA,
